Extract DumpNodes to a new debug_util library.
This change moves the `DumpNodes` function from `immutable_converter_main.cc` into a new library, `debug_util`. This will reduce code duplication.

#codehealth

PiperOrigin-RevId: 907966727
