+++ src/converter/immutable_converter_main.cc
@@ -57,26 +57,20 @@ id	key	value	begin_pos	end_pos	lid	rid	wcost	cost	bnext	enext	prev	next
 */
 // clang-format on
 
-#include <cstddef>
-#include <cstdint>
 #include <iostream>
 #include <memory>
 #include <ostream>
 #include <string>
 #include <utility>
-#include <vector>
 
-#include "absl/container/flat_hash_map.h"
 #include "absl/flags/flag.h"
-#include "absl/strings/str_cat.h"
-#include "absl/strings/str_join.h"
+#include "absl/status/statusor.h"
 #include "absl/strings/string_view.h"
-#include "absl/types/span.h"
 #include "base/file_stream.h"
 #include "base/init_mozc.h"
+#include "converter/debug_util.h"
 #include "converter/immutable_converter.h"
 #include "converter/lattice.h"
-#include "converter/node.h"
 #include "converter/segments.h"
 #include "data_manager/data_manager.h"
 #include "data_manager/oss/oss_data_manager.h"
@@ -89,112 +83,8 @@ ABSL_FLAG(std::string, dictionary, "", "Dictionary: 'oss' or 'test'");
 ABSL_FLAG(std::string, output, "", "Output file");
 
 namespace mozc {
+namespace converter {
 namespace {
-
-std::string DumpNodes(const Lattice& lattice) {
-  // Map from Node pointer to a unique ID.
-  absl::flat_hash_map<const Node*, uint32_t> node_id_map;
-  node_id_map[nullptr] = 0;
-
-  // Returns a unique ID for the given node.
-  auto node_id = [&node_id_map](const Node* node) -> uint32_t {
-    auto it = node_id_map.find(node);
-    if (it != node_id_map.end()) {
-      return it->second;
-    }
-    uint32_t id = node_id_map.size();
-    node_id_map[node] = id;
-    return id;
-  };
-
-  auto bnext_id = [&](const Node& node) -> uint32_t {
-    const absl::Span<const Node* const>& begin_nodes =
-        lattice.begin_nodes(node.begin_pos);
-    // If the node is a POS node, returns the first begin node.
-    if (node.key.empty() && node.value == "POS") {
-      return node_id(begin_nodes[0]);
-    }
-    for (size_t i = 0; i < begin_nodes.size() - 1; ++i) {
-      if (begin_nodes[i] == &node) {
-        return node_id(begin_nodes[i + 1]);
-      }
-    }
-    return 0;
-  };
-
-  auto enext_id = [&](const Node& node) -> uint32_t {
-    const absl::Span<const Node* const>& end_nodes =
-        lattice.end_nodes(node.end_pos);
-    for (size_t i = 1; i < end_nodes.size(); ++i) {
-      if (end_nodes[i] == &node) {
-        return node_id(end_nodes[i - 1]);
-      }
-    }
-    return 0;
-  };
-
-  // Returns a string representation of the given node in TSV.
-  auto dump_node = [&](const Node& node) -> std::string {
-    return absl::StrCat(node_id(&node), "\t", node.key, "\t", node.value, "\t",
-                        node.begin_pos, "\t", node.end_pos, "\t", node.lid,
-                        "\t", node.rid, "\t", node.wcost, "\t", node.cost, "\t",
-                        bnext_id(node), "\t", enext_id(node), "\t",
-                        node_id(node.prev), "\t", node_id(node.next), "\n");
-  };
-
-  std::string output;
-
-  // Output header.
-  output =
-      absl::StrJoin({"id", "key", "value", "begin_pos", "end_pos", "lid", "rid",
-                     "wcost", "cost", "bnext", "enext", "prev", "next"},
-                    "\t");
-  absl::StrAppend(&output, "\n");
-
-  // Output BOS node.
-  absl::StrAppend(&output, dump_node(*lattice.bos_node()));
-
-  // Make position nodes.
-  std::vector<Node> pos_nodes;
-  for (size_t i = 0; i <= lattice.key().size(); ++i) {
-    if (lattice.begin_nodes(i).empty()) {
-      continue;
-    }
-
-    Node pos_node;
-    pos_node.begin_pos = i;
-    pos_node.end_pos = i;
-    pos_node.value = "POS";
-    pos_nodes.push_back(std::move(pos_node));
-  }
-  for (size_t i = 0; i < pos_nodes.size() - 1; ++i) {
-    pos_nodes[i].next = &pos_nodes[i + 1];
-  }
-
-  // Output position nodes.
-  for (const Node& pos_node : pos_nodes) {
-    absl::StrAppend(&output, dump_node(pos_node));
-  }
-
-  // Output nodes.
-  for (size_t i = 0; i <= lattice.key().size(); ++i) {
-    if (lattice.begin_nodes(i).empty()) {
-      continue;
-    }
-    {
-      Node pos_node;
-      pos_node.begin_pos = i;
-      pos_node.end_pos = i;
-      pos_node.value = "POS";
-      pos_nodes.push_back(std::move(pos_node));
-    }
-
-    for (const Node* node : lattice.begin_nodes(i)) {
-      absl::StrAppend(&output, dump_node(*node));
-    }
-  }
-  return output;
-}
 
 bool ExecCommand(const ImmutableConverter& immutable_converter,
                  absl::string_view query, absl::string_view output) {
@@ -235,24 +125,33 @@ std::unique_ptr<const DataManager> CreateDataManager(
 
   return std::make_unique<const oss::OssDataManager>();
 }
-
 }  // namespace
-}  // namespace mozc
 
-int main(int argc, char** argv) {
-  mozc::InitMozc(argv[0], &argc, &argv);
+int RunMain(int argc, char** argv) {
+  InitMozc(argv[0], &argc, &argv);
 
-  std::unique_ptr<const mozc::DataManager> data_manager =
-      mozc::CreateDataManager(absl::GetFlag(FLAGS_dictionary));
-
-  std::unique_ptr<mozc::engine::Modules> modules =
-      mozc::engine::Modules::Create(std::move(data_manager)).value();
-  mozc::ImmutableConverter immutable_converter(*modules);
+  std::unique_ptr<const DataManager> data_manager =
+      CreateDataManager(absl::GetFlag(FLAGS_dictionary));
+  absl::StatusOr<std::unique_ptr<engine::Modules>> modules_opt =
+      engine::Modules::Create(std::move(data_manager));
+  if (!modules_opt.ok()) {
+    std::cerr << "Failed to create modules for "
+              << absl::GetFlag(FLAGS_dictionary) << std::endl;
+    return 1;
+  }
+  std::unique_ptr<engine::Modules> modules = std::move(modules_opt.value());
+  ImmutableConverter immutable_converter(*modules);
 
   std::string query = absl::GetFlag(FLAGS_query);
   std::string output = absl::GetFlag(FLAGS_output);
-  if (!mozc::ExecCommand(immutable_converter, query, output)) {
+  if (!ExecCommand(immutable_converter, query, output)) {
     std::cerr << "ExecCommand() return false" << std::endl;
+    return 1;
   }
   return 0;
 }
+
+}  // namespace converter
+}  // namespace mozc
+
+int main(int argc, char** argv) { return mozc::converter::RunMain(argc, argv); }
